Move IO file system tests into subfolder.
