+++ source/Atmoos.World.IO.Test/FileSystem/CurrentFileSystemTest.cs
@@ -1,6 +1,6 @@
 using Atmoos.World.IO.FileSystem;
 
-namespace Atmoos.World.IO.Test;
+namespace Atmoos.World.IO.Test.FileSystem;
 
 public sealed class CurrentFileSystemTest
 {
